Move derive address tests to their own file
We also take the opportunity to reformat the tests into tabular tests.

https://github.com/keep-network/keep-core/blob/master/docs/development/go-guidelines.adoc#testing
